% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Function is run to clean data from NOAA and transforms to prescribed form}
\usage{
eq_clean_data(dataframe)
}
\arguments{
\item{dataframe}{Dataframe of loaded NOAA data}
}
\description{
This function used to clean data from NOAA. File is loaded to dataframe from
TSV file in gz file. first part of function is used to create bce variable,
which is used to determine which function will be used to transform to data.
In case that there is a negative sign then parse_bce_ymd will be used to transform
date variable. YEAR, MONTH and DAY will be transformed to string character and padded
to have format YYYY MM DD. Creation of date is dependent on bce variable. If bce
variable is TRUE then parse_bce_ymd function will be used.
Final part of function is used to transform LATITUDE and LONGITUDE to numerical
format.
}
\examples{
\dontrun{
filename<-system.file("data","earthquakes.tsv.gz",package="ProjectCapstone")
dataframe <- read_delim(filename)
eq_clean_data(dataframe)
}

}
