% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleaning variable LOCATION_NAME from inclusion of COUNTRY}
\usage{
eq_location_clean(dataframe)
}
\arguments{
\item{dataframe}{Dataframe to be transformed}
}
\description{
Function will be used to transform column LOCATION_NAME to clean form.
First part of process is removing COUNTRY from LOCATION_NAME with function 
str_remove_all.
Secund part is transforming LOCATION_NAME to title caps.
Final part of function str_remove_all is used to replace : and ; 
in LOCATION_NAME to clean string.
}
\examples{
\dontrun{
filename<-system.file("data","earthquakes.tsv.gz",package="ProjectCapstone")
dataframe <- read_delim(filename)
dataframe <- eq_clean_data(dataframe)
eq_clean_data(dataframe)
}

}
