% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_proto.R
\docType{data}
\name{GeomTimeLineAnnotation}
\alias{GeomTimeLineAnnotation}
\title{Funcion for annotating data for eartquake data from NOAA.}
\format{
An object of class \code{GeomTimeLineAnnotation} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomTimeLineAnnotation
}
\description{
Funcion for annotating data for eartquake data from NOAA.
}
\keyword{datasets}
